Added extended Geometry with a feild that is the DrawableSceneComponent associated with it.

+++ GameEngine/Levels/Characters/Bunny.cs
@@ -44,7 +44,7 @@ namespace Gdd.Game.Engine.Levels.Characters
         /// <summary>
         /// The life.
         /// </summary>
-        private float life;
+        private int life;
 
         #endregion
